here's the deal...
...I'm just gonna clone it because no amount of bigbrain-energy-;)-begat-by idk y'all -- the i3 tiling window manager and like, a well configured workflow, 
can overcome the awfulness of the github web editor
y'all[^1]. especially when footnotes are involved?

[^1]: I'm a y'all transplant but I'm putting this 'footnote' here out of curiosity -- will github treat the "Extended description" as plaintext?

I've noticed over time in a lot of web apps that you can throw in some markdown syntax and end up with this very interesting 1/3-markdown-1/3-hypertext-1/3-plaintext rendering situation (sometimes with emoji to boot ;) <- what did it do y'all? ) and it's honestly comedic, but sometimes I pause and realize it's rather concerning?