oh dear! I'm supposed to be writing a provocation....! am I?
it's funny, arbitrary design choices, even in prog. lang.'s. Like here in markdown: I did not check the gh preview but I am willing to wager (I love the alliteration of that btw, I won't admit it in very public writing, but man 'willing to wager' just rolls off the tongue don't it? ::: if you haven't gathered the "Extended description" of an arbitrary git commit on gh is not easily findable, not that these tidnibbles ;)  are *hidden*) 

THAT (dud, try to backtrace that, you donut get it I have a thread in mind when I open my mouth [sic.: fingers].), that that that (pay attention because I do this in spoken word too, idiot.) THAT;

that you have no idea what a person is saying from the words that sound. 

does any of this make sense? No. Language is useless.

:::

;'essebtie  est invisib;q ppur les yeux 


I wouldn't be bothering you, father


don't read into any of that; it isn't for you if you haven;t gotten it



that even though I didn't check the preview that this one whitespace will insert the blank line to, 
deliver, 
what I had to say

are you starting to get it?
surely not, because language is useless
and our days 

our

numbered and we may as well use our yeux
or eyebrows even, they evince more

instead